Add exporatory notebook
What works:
 * quantile normalization and computing the correlation matrix
 * "deconvolution" assigning weights to each single cell
 * statistical test of association of each cell based on these weights

What does not work
 * LASSO regression based on correlation (i.e. scissor algorithm). The
 graph-based penalty is not trivial to implement (I found no off-the-shelf
 solutions). Using vanilla LASSO regression does not yield any positive
 weights, using ridge results in no signal (at least for LUAD TP53).

EDIT: I accidentically  screwed it up with a wrong "jupytext sync"
and my version is lost now.

Anyway, the value is in its learning, and if I redo this, I can
very well implement it properly as a python library. With the
proper scikit-learn functions, the exporation was less
than an afternoon anyway.